Initial commit
Warning: This content, especially the code will be heavily subject to change in the future. Even the implementation language will probably change!

Tribes: gathering is a multiplayer game where players try to survive and thrive as tribe and improve their position in it. As the players help the tribe overcome various challenges they contribute to the unique story of their tribe.

While many other society-building games focus on technological, infrastructural and/or character building, the main theme of this game is the exploring of social roles, social standing and emergent social behaviour inside a community of humans.

Currently main content is the "tribes-gathering" textual context file and Java-based exploratory core concept model (draft)